requirements.txt
# requirements.txt for HealthBridge AI RAG Chatbot

# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# RAG & AI
langchain==0.1.17
langchain-community==0.0.38
langchain-groq==0.1.5
groq==0.9.0

# Vector Databases
qdrant-client==1.9.1

# Document Processing
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1

# Embeddings
sentence-transformers==2.7.0
transformers==4.41.2

# Utils
python-dotenv==1.0.0
